requirements.txt
torch
torchvision
torchsummary
tqdm
matplotlib
Pillow